Checkpoint2

+++ driver/driver.cpp
@@ -6,14 +6,14 @@ int TDriver::parse(const std::string& f) {
     file = f;
     location.initialize(&file);
     scan_begin();
-    parser.set_debug_level(trace_parsing);
-    int parse_result = parser();
+    parser.set_debug_level(static_cast<int>(trace_parsing));
+    int parse_result = static_cast<int>(parser());
     scan_end();
     return parse_result;
 }
 
 void TDriver::scan_begin() {
-    scanner.set_debug(trace_scanning);
+    scanner.set_debug(static_cast<int>(trace_scanning));
     if (file.empty() || file == "-") {
     } else {
         stream.open(file);
